tiles back to integer math
further, still incomplete, work on blocking at borders

git-svn-id: http://fritzing.googlecode.com/svn/trunk@4617 85f94c43-f63a-0410-9d22-21439860c375

+++ fritzing/src/autoroute/jrouter.h
@@ -81,8 +81,8 @@ struct SeedTree {
 		Down
 	};
 	
-	qreal restrictionMin;
-	qreal restrictionMax;
+	int restrictionMin;
+	int restrictionMax;
 	bool restricted;
 	struct Tile * seed;
 	SeedTree * parent;
